maintain: Spliting all code and logic in services and models in difer… (#143)
* maintain: Spliting all code and logic in services and models in diferent apps.

* maintain: cleaning.

* maintain: cleaning.
