+++ apps/analytics/tests.py
