+++ apps/contents/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ContentsConfig(AppConfig):
+    name = 'apps.contents'
