+++ apps/contents/services.py
@@ -1,0 +1,177 @@
+import logging
+import os
+from concurrent.futures import ThreadPoolExecutor
+
+import requests
+from dotenv import load_dotenv
+
+# Load environment variables
+load_dotenv()
+
+# Configure logging
+# logger = logging.getLogger(__name__)
+
+urls_list = os.getenv('API_BASE_URLS', '').split(',')
+keys_list = os.getenv('API_KEYS_DJANGO', '').split(',')
+API_CONFIG = dict(zip(urls_list, keys_list))
+
+TMDB_API_KEY = os.getenv('TMDB_API_KEY')
+
+OPTIONS = {
+    'plataformas': ['CinePlus', 'StreamHub', 'PlayMax'],
+    'idiomas': ['Català', 'Castellano', 'English', 'Français']
+}
+
+def get_tmdb_image(title):
+    try:
+        response = requests.get(
+            "https://api.themoviedb.org/3/search/multi",
+            params={
+                "api_key": TMDB_API_KEY,
+                "query": title,
+                "language": "en"
+            },
+            timeout=2
+        )
+        if response.status_code == 200:
+            results = response.json().get("results", [])
+            if results and results[0].get("poster_path"):
+                return f"https://image.tmdb.org/t/p/w500{results[0]['poster_path']}"
+    except (requests.RequestException, ValueError):
+        pass
+    return 'https://via.placeholder.com/300x450'
+
+
+def enrich_tmdb_images(items):
+    def load_image(item):
+        item['imatge'] = get_tmdb_image(item['titol'])
+        return item
+
+    with ThreadPoolExecutor(max_workers=10) as executor:
+        items = list(executor.map(load_image, items))
+    return items
+
+
+# --- 3. DATA MAPPING ---
+
+def map_data(item, port):
+    platforms = {"8080": "CinePlus", "8081": "StreamHub", "8082": "PlayMax"}
+    port_net = str(port).replace('/','')
+
+
+    title = item.get('title') or item.get('titol') or "Sense títol"
+    synopsis = item.get('synopsis') or "Sense sinopsi disponible."
+    content_year = item.get('year') or item.get('start_year') or 0
+
+    return {
+        'id': f"{port}_{item.get('id')}",
+        'titol': title,
+        'sinopsi': synopsis,
+        'any': content_year,
+        'any_fi': item.get('end_year'),
+        'total_seasons': item.get('total_seasons'),
+        'rating': item.get('rating', '0.0'),
+        'imatge': item.get('imatge') or 'https://via.placeholder.com/300x450',
+        'plataforma': platforms.get(port, "Altres"),
+        'genre_id': item.get('genre_id'),
+        'director_id': item.get('director_id'),
+        'age_rating_id': item.get('age_rating_id'),
+        'genere_nom': "General",
+        'director_nom': "Desconegut",
+        'edat_nom': "N/A"
+    }
+
+def deduplicate_content(llista):
+    vistos = {}
+    for item in llista:
+        titol = item['titol'].strip().lower()
+        if titol not in vistos:
+            item['plataformes_disponibles'] = [item['plataforma']]
+            vistos[titol] = item
+        else:
+            if item['plataforma'] not in vistos[titol]['plataformes_disponibles']:
+                vistos[titol]['plataformes_disponibles'].append(item['plataforma'])
+    return list(vistos.values())
+
+# --- 4. STREAMSYNC API CALLS ---
+
+def get_all_movies(query=None):
+    results = []
+    for base_url, key in API_CONFIG.items():
+        headers = {'x-api-key': key}
+        port = base_url.split(':')[-1]
+        params = {'title': query} if query else {}
+        try:
+            response = requests.get(f"{base_url}/movies", headers=headers, params=params, timeout=2)
+            if response.status_code == 200:
+                for item in response.json():
+                    obj = map_data(item, port)
+                    obj['tipus'] = 'movie'
+                    results.append(obj)
+        except:
+            pass
+    return deduplicate_content(results)  # ✅ Deduplicació
+
+
+def enrich_api_data(content_list):
+    genres_api = get_genres_from_api()
+    ratings_api = get_age_ratings_from_api()
+
+    genre_map = {str(g['id']): g['name'] for g in genres_api}
+    rating_map = {str(r['id']): r.get('description', 'N/A') for r in ratings_api}
+
+    for item in content_list:
+        gid = str(item.get('genre_id'))
+        eid = str(item.get('age_rating_id'))
+        item['genere_nom'] = genre_map.get(gid, "General")
+        item['edat_nom'] = rating_map.get(eid, "N/A")
+        if 'tipus' not in item:
+            item['tipus'] = item.get('media_type', 'movie')
+
+    return content_list
+
+
+def get_all_series(query=None):
+    results = []
+    for base_url, key in API_CONFIG.items():
+        headers = {'x-api-key': key}
+        port = base_url.split(':')[-1]
+        params = {'title': query} if query else {}
+        try:
+            response = requests.get(f"{base_url}/series", headers=headers, params=params, timeout=2)
+            if response.status_code == 200:
+                for item in response.json():
+                    obj = map_data(item, port)
+                    obj['tipus'] = 'series'
+                    results.append(obj)
+        except:
+            pass
+    return deduplicate_content(results)  # ✅ Deduplicació
+
+
+def get_genres_from_api():
+    for base_url, key in API_CONFIG.items():
+        try:
+            return requests.get(f"{base_url}/genres", headers={'x-api-key': key}, timeout=1).json()
+        except:
+            continue
+    return []
+
+
+def get_directors_from_api():
+    for base_url, key in API_CONFIG.items():
+        try:
+            return requests.get(f"{base_url}/directors", headers={'x-api-key': key}, timeout=1).json()
+        except:
+            continue
+    return []
+
+
+def get_age_ratings_from_api():
+    for base_url, key in API_CONFIG.items():
+        try:
+            return requests.get(f"{base_url}/age-ratings", headers={'x-api-key': key}, timeout=1).json()
+        except:
+            continue
+    return []
+
