+++ apps/reviews/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ReviewsConfig(AppConfig):
+    name = 'apps.reviews'
