+++ apps/users/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import Profile
+from apps.users.models.models import Profile
 
 @admin.register(Profile)
 class ProfileAdmin(admin.ModelAdmin):
