+++ requirements.txt
@@ -37,4 +37,6 @@ wheel==0.42.0
 djangorestframework~=3.15.1
 python-decouple~=3.8
 
+django-cookie-consent~=1.0.0
+
 textstat
